Persist weights after training and load for prediction
